% Auto-generated: do not edit by hand
\name{''DashMonacoEditor}

\alias{''DashMonacoEditor}

\title{DashMonacoEditor component}

\description{

}

\usage{
''DashMonacoEditor(id=NULL, height=NULL, language=NULL, theme=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Character. Height of the editor}

\item{language}{Character. The language of the editor}

\item{theme}{Character. current theme "light","vs-dark"}

\item{value}{Character. current code}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

