% Auto-generated: do not edit by hand
\name{dashMonacoEditor}

\alias{dashMonacoEditor}

\title{DashMonacoEditor component}

\description{

}

\usage{
dashMonacoEditor(id=NULL, code=NULL, defaultLanguage=NULL, height=NULL,
language=NULL, options=NULL, theme=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{code}{Character. }

\item{defaultLanguage}{Character. }

\item{height}{Character. }

\item{language}{Character. }

\item{options}{Named list. }

\item{theme}{Character. }

\item{width}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

